+++ CMakeLists.txt
@@ -1,11 +1,5 @@
 ﻿cmake_minimum_required (VERSION 3.28)
 
-# Enable Hot Reload for MSVC compilers if supported.
-if (POLICY CMP0141)
-  cmake_policy(SET CMP0141 NEW)
-  set(CMAKE_MSVC_DEBUG_INFORMATION_FORMAT "$<IF:$<AND:$<C_COMPILER_ID:MSVC>,$<CXX_COMPILER_ID:MSVC>>,$<$<CONFIG:Debug,RelWithDebInfo>:EditAndContinue>,$<$<CONFIG:Debug,RelWithDebInfo>:ProgramDatabase>>")
-endif()
-
 project(
   NavKit 
   VERSION 0.1
@@ -24,5 +18,4 @@ find_package(GLUT REQUIRED)
 find_package(GLU REQUIRED)
 
 add_subdirectory(src)
-# include_directories( ${OPENGL_INCLUDE_DIRS}  ${GLUT_INCLUDE_DIRS} )
 #FIND_LIBRARY(NAVWEAKNESS_LIBRARIES NAMES navweakness NavWeakness.lib PATHS "External/NavWeakness/NavWeakness.lib")
