+++ include/ResourceLib_HM3/ResourceConverter.h
