+++ include/ResourceLib_HM3/ResourceGenerator.h
