+++ include/ResourceLib_HM3/ResourceLib.h
