+++ include/ResourceLib_HM3/ResourceLibCommon.h
