+++ include/ResourceLib_HM3/ResourceLib_HM3.h
