+++ include/ResourceLib_HM3/Util/BinaryStreamReader.h
