+++ include/ResourceLib_HM3/Util/BinaryStreamWriter.h
