+++ include/ResourceLib_HM3/Util/PortableIntrinsics.h
