+++ include/ResourceLib_HM3/ZHM/Hash.h
