+++ include/ResourceLib_HM3/ZHM/ZEncryptedString.h
