+++ include/ResourceLib_HM3/ZHM/ZHMCustomProperties.h
