+++ include/ResourceLib_HM3/ZHM/ZHMCustomTypes.h
