+++ include/ResourceLib_HM3/ZHM/ZHMEquality.h
