+++ include/ResourceLib_HM3/ZHM/ZHMPrimitives.h
