+++ include/ResourceLib_HM3/ZHM/ZHMPtr.h
