+++ include/ResourceLib_HM3/ZHM/ZHMSerializer.h
