+++ include/ResourceLib_HM3/ZHM/ZHMTypeInfo.h
