+++ include/ResourceLib_HM3/ZHM/ZString.h
