+++ include/ResourceLib_HM3/ZHM/ZVariant.h
